requirements.txt
bcrypt==4.0.1
blinker==1.5
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
Deprecated==1.2.13
dnspython==2.2.1
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.1.0
Flask-Minify==0.41
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
lesscpy==0.15.1
limits==2.8.0
MarkupSafe==2.1.1
ordered-set==4.1.0
packaging==22.0
ply==3.11
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.0
rcssmin==1.1.1
requests==2.28.1
rich==12.6.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
xxhash==3.1.0
